restructured the interface to math new develop convention
