--- src/Gold.Redis/Gold.Redis.Common/Interfaces/Db/IRedisCommandsExecutor.cs
@@ -1,10 +1,0 @@
-﻿using Gold.Redis.Common.Models.Commands;
-using System.Threading.Tasks;
-
-namespace Gold.Redis.Common.Interfaces.Db
-{
-    public interface IRedisCommandExecutor
-    {
-        Task<T> Execute<T>(Command command);
-    }
-}
