+++ src/Gold.Redis/Gold.Redis.HighLevelClient/Db/RedisCommandsExecutor.cs
@@ -1,10 +1,10 @@
-﻿using Gold.Redis.Common.Interfaces.Db;
-using Gold.Redis.Common.Models.Commands;
+﻿using Gold.Redis.Common.Models.Commands;
 using Gold.Redis.LowLevelClient.Interfaces;
 using Newtonsoft.Json;
 using System;
 using System.IO;
 using System.Threading.Tasks;
+using Gold.Redis.HighLevelClient.Interfaces;
 
 namespace Gold.Redis.HighLevelClient.Db
 {
