+++ src/Gold.Redis/Gold.Redis.HighLevelClient/Interfaces/IRedisCommandsExecutor.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using Gold.Redis.Common.Models.Commands;
+
+namespace Gold.Redis.HighLevelClient.Interfaces
+{
+    public interface IRedisCommandExecutor
+    {
+        Task<T> Execute<T>(Command command);
+    }
+}
