+++ src/Gold.Redis/Gold.Redis.HighLevelClient/Interfaces/IRedisKeysDb.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace Gold.Redis.Common.Interfaces.Db
+namespace Gold.Redis.HighLevelClient.Interfaces
 {
     public interface IRedisKeysDb : IRedisDbGeneralOperations
     {
